assistant/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.6.2.post1
attrs==24.3.0
av==14.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dependency-injector==4.41.0
distro==1.9.0
fastapi==0.97.0
fastapi-versioning==0.10.0
flatbuffers==25.1.21
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.8.2
livekit==0.19.1
livekit-agents==0.12.8
livekit-api==0.8.1
livekit-plugins-openai==0.10.15
livekit-plugins-silero==0.7.4
livekit-protocol==0.8.0
mpmath==1.3.0
multidict==6.1.0
numpy==2.2.2
onnxruntime==1.20.1
openai==1.59.8
packaging==24.2
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pycares==4.5.0
pycparser==2.22
pydantic==1.10.19
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.18
redis==5.2.1
requests==2.32.3
sh==2.0.4
six==1.16.0
sniffio==1.3.1
sse-starlette==1.6.1
starlette==0.27.0
sympy==1.13.3
tenacity==9.0.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
yarl==1.18.3
